cffi: call libffi through a jit_libffi port
Port `rpython/rlib/jit_libffi.py` as `_cffi_backend/jit_libffi.rs`: the
`CIF_DESCRIPTION` header accessors, the `types` kind/size readers, and the
result-kind split that ends in the four `libffi_call` oopspec leaves and
`jit_ffi_call_impl_any`.  `ctypefunc::do_call` calls `jit_ffi_call` instead
of invoking libffi itself, and `ctypefunc::cif` keeps only the builder.

`W_CTypeFunc.cif_descr` and `W_FunctionWrapper.{fnptr,directfnptr}` are
`usize` rather than `*mut u8`.  `W_CType`, `W_FunctionWrapper` and
`W_RawFuncType` declare their immutable fields, and
`W_FunctionWrapper.descr_call` promotes the raw function type.

`jit_fnaddr` registers the `jit_libffi` readers and oopspec leaves in place
of the removed `ctypefunc::cif` helpers.

The four `assert_no_libffi_call_oopspec` panics in `JitCodeMachine` are
removed: that layer answers upstream's `None  # cannot be handled by
direct_libffi_call()`.

+++ pyre/pyre-interpreter/src/module/_cffi_backend/cdataobj.rs
@@ -926,7 +926,7 @@ pub fn __majit_wrap_cdata_call(args: &[PyObjectRef]) -> Result<PyObjectRef, PyEr
     let cdata = cdata_arg(args[0])?;
     let ct = cdata.ctype_ref()?;
     if ct.kind == ctypeobj::KIND_FUNC {
-        return super::ctypefunc::call(ct, cdata.ptr, &args[1..]);
+        return super::ctypefunc::call(ct, cdata.ptr as usize, &args[1..]);
     }
     Err(PyError::type_error(format!(
         "cdata '{}' is not callable",
